Fixed deprecated use of platform.linux_distribution.
This is a correction for the deprecated usage of platform.linux_distribution, which is not available on Python3.8, or above.
To fix this, the distro package is used. Distro package is only available for Linux and BSD systems by this moment, so this non-elegant solution was used to not ask for this package in other systems.